Release 10.47.43
git-svn-id: http://svn.code.sf.net/p/netpbm/code/stable@1852 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ converter/pbm/cmuwmtopbm.c
@@ -40,7 +40,7 @@ readCmuwmHeader(FILE *         const ifP,
     rc = pm_readbiglong(ifP, &l);
     if (rc == -1 )
         pm_error(initReadError);
-    if (l != CMUWM_MAGIC)
+    if ((uint32_t)l != CMUWM_MAGIC)
         pm_error("bad magic number in CMU window manager file");
     rc = pm_readbiglong(ifP, &l);
     if (rc == -1)
